requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==3.2.12
django-bootstrap5==22.2
Pillow==9.3.0
pytz==2022.6
soupsieve==2.3.2.post1
sqlparse==0.4.3
